+++ agb/examples/text_render.rs
@@ -3,6 +3,7 @@
 
 use agb::{
     display::{
+        palette16::Palette16,
         tiled::{
             DynamicTile, RegularBackgroundSize, RegularBackgroundTiles, TileFormat, VRAM_MANAGER,
         },
@@ -20,10 +21,13 @@ fn main(mut gba: agb::Gba) -> ! {
     let mut gfx = gba.display.graphics.get();
     let vblank = agb::interrupt::VBlank::get();
 
-    VRAM_MANAGER.set_background_palette_raw(&[
-        0x0000, 0x0ff0, 0x00ff, 0xf00f, 0xf0f0, 0x0f0f, 0xaaaa, 0x5555, 0x0000, 0x0000, 0x0000,
-        0x0000, 0x0000, 0x0000, 0x0000, 0x0000,
-    ]);
+    VRAM_MANAGER.set_background_palette(
+        0,
+        &Palette16::new([
+            0x0000, 0x0ff0, 0x00ff, 0xf00f, 0xf0f0, 0x0f0f, 0xaaaa, 0x5555, 0x0000, 0x0000, 0x0000,
+            0x0000, 0x0000, 0x0000, 0x0000, 0x0000,
+        ]),
+    );
 
     let background_tile = DynamicTile::new().fill_with(0);